# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PayAgent Escrow API — Celo cUSD smart contract operations
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  details?: string;
}

export interface DepositEscrowRequest {
  /** Unique job identifier (hex string, will be hashed to bytes32) */
  jobId: string;
  /** Wallet address of the gig worker */
  workerAddress: string;
  /** Amount of cUSD to escrow (in human-readable units e.g. "10.5") */
  amount: string;
  /** Human-readable job title for receipt generation */
  jobTitle: string;
}

export interface ReleasePaymentRequest {
  /** The job ID to release payment for (hex bytes32) */
  jobId: string;
  /** Optional proof of completion (URL, hash, or description) */
  completionProof?: string;
}

export interface DisputeRequest {
  /** The job ID to dispute */
  jobId: string;
  /** Address of the party initiating the dispute (worker or employer) */
  initiatorAddress: string;
  /** Reason for the dispute */
  reason?: string;
}

/**
 * Target network (celo = mainnet, alfajores = testnet)
 */
export type DeployRequestNetwork =
  (typeof DeployRequestNetwork)[keyof typeof DeployRequestNetwork];

export const DeployRequestNetwork = {
  celo: "celo",
  alfajores: "alfajores",
} as const;

export interface DeployRequest {
  /** Target network (celo = mainnet, alfajores = testnet) */
  network: DeployRequestNetwork;
  /** Override cUSD token address (optional, uses well-known defaults) */
  cUSDAddress?: string;
}

export type TransactionResultStatus =
  (typeof TransactionResultStatus)[keyof typeof TransactionResultStatus];

export const TransactionResultStatus = {
  success: "success",
  pending: "pending",
  failed: "failed",
} as const;

export interface TransactionResult {
  txHash: string;
  blockNumber?: number;
  status: TransactionResultStatus;
  jobId?: string;
  gasUsed?: string;
  timestamp?: number;
}

export type EscrowDetailsStatus =
  (typeof EscrowDetailsStatus)[keyof typeof EscrowDetailsStatus];

export const EscrowDetailsStatus = {
  None: "None",
  Escrowed: "Escrowed",
  Released: "Released",
  Disputed: "Disputed",
  Resolved: "Resolved",
  Refunded: "Refunded",
} as const;

export interface EscrowDetails {
  jobId: string;
  employer: string;
  worker: string;
  /** Net amount in cUSD (human-readable) */
  amount: string;
  /** Platform fee in cUSD (human-readable) */
  fee: string;
  status: EscrowDetailsStatus;
  /** Unix timestamp */
  createdAt: number;
  jobTitle: string;
}

export interface JobIdList {
  address: string;
  jobIds: string[];
  count: number;
}

export type PaymentReceiptStatus =
  (typeof PaymentReceiptStatus)[keyof typeof PaymentReceiptStatus];

export const PaymentReceiptStatus = {
  Released: "Released",
  Resolved: "Resolved",
} as const;

export interface PaymentReceipt {
  jobId: string;
  jobTitle: string;
  worker: string;
  employer: string;
  /** Amount paid in cUSD */
  amount: string;
  /** Release transaction hash */
  txHash: string;
  blockNumber: number;
  network: string;
  contractAddress: string;
  timestamp: number;
  /** Link to Celoscan explorer for the transaction */
  celoscanUrl: string;
  status: PaymentReceiptStatus;
}

export interface ReceiptList {
  worker: string;
  receipts: PaymentReceipt[];
  count: number;
}

export interface DeployResult {
  contractAddress: string;
  txHash: string;
  network: string;
  deployerAddress: string;
  aiAgentAddress: string;
  cUSDAddress: string;
  blockNumber: number;
  celoscanUrl: string;
}

export interface ContractStatus {
  deployed: boolean;
  contractAddress?: string;
  network?: string;
  aiAgent?: string;
  feeBps?: number;
  collectedFees?: string;
  deployerAddress?: string;
  celoscanUrl?: string;
}

export type TransactionEntryType =
  (typeof TransactionEntryType)[keyof typeof TransactionEntryType];

export const TransactionEntryType = {
  deposit: "deposit",
  release: "release",
  dispute: "dispute",
  resolve: "resolve",
  refund: "refund",
  deploy: "deploy",
} as const;

export interface TransactionEntry {
  id: number;
  type: TransactionEntryType;
  jobId?: string;
  txHash: string;
  from?: string;
  to?: string;
  amount?: string;
  status: string;
  blockNumber?: number;
  network?: string;
  createdAt: number;
}

export interface TransactionLog {
  transactions: TransactionEntry[];
  total: number;
  limit: number;
  offset: number;
}

export type GetTransactionLogParams = {
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
  /**
   * @minimum 0
   */
  offset?: number;
};