# lib/api-zod/src/generated/types/deployRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PayAgent Escrow API — Celo cUSD smart contract operations
 * OpenAPI spec version: 0.1.0
 */
import type { DeployRequestNetwork } from "./deployRequestNetwork";

export interface DeployRequest {
  /** Target network (celo = mainnet, alfajores = testnet) */
  network: DeployRequestNetwork;
  /** Override cUSD token address (optional, uses well-known defaults) */
  cUSDAddress?: string;
}