# lib/api-zod/src/generated/types/disputeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PayAgent Escrow API — Celo cUSD smart contract operations
 * OpenAPI spec version: 0.1.0
 */

export interface DisputeRequest {
  /** The job ID to dispute */
  jobId: string;
  /** Address of the party initiating the dispute (worker or employer) */
  initiatorAddress: string;
  /** Reason for the dispute */
  reason?: string;
}