# lib/api-zod/src/generated/types/paymentReceipt.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PayAgent Escrow API — Celo cUSD smart contract operations
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentReceiptStatus } from "./paymentReceiptStatus";

export interface PaymentReceipt {
  jobId: string;
  jobTitle: string;
  worker: string;
  employer: string;
  /** Amount paid in cUSD */
  amount: string;
  /** Release transaction hash */
  txHash: string;
  blockNumber: number;
  network: string;
  contractAddress: string;
  timestamp: number;
  /** Link to Celoscan explorer for the transaction */
  celoscanUrl: string;
  status: PaymentReceiptStatus;
}