# lib/api-zod/src/generated/types/receiptList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PayAgent Escrow API — Celo cUSD smart contract operations
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentReceipt } from "./paymentReceipt";

export interface ReceiptList {
  worker: string;
  receipts: PaymentReceipt[];
  count: number;
}