# lib/api-zod/src/generated/types/releasePaymentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PayAgent Escrow API — Celo cUSD smart contract operations
 * OpenAPI spec version: 0.1.0
 */

export interface ReleasePaymentRequest {
  /** The job ID to release payment for (hex bytes32) */
  jobId: string;
  /** Optional proof of completion (URL, hash, or description) */
  completionProof?: string;
}