# lib/api-zod/src/generated/types/transactionEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PayAgent Escrow API — Celo cUSD smart contract operations
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionEntryType } from "./transactionEntryType";

export interface TransactionEntry {
  id: number;
  type: TransactionEntryType;
  jobId?: string;
  txHash: string;
  from?: string;
  to?: string;
  amount?: string;
  status: string;
  blockNumber?: number;
  network?: string;
  createdAt: number;
}