# lib/api-zod/src/generated/types/transactionResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PayAgent Escrow API — Celo cUSD smart contract operations
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionResultStatus } from "./transactionResultStatus";

export interface TransactionResult {
  txHash: string;
  blockNumber?: number;
  status: TransactionResultStatus;
  jobId?: string;
  gasUsed?: string;
  timestamp?: number;
}